refactor(move): move `errorMsgTag` function from `src/utils/nunjucks.ts` to `src/utils/index.ts`
The errorMsgTag function is not dedicated to Nunjucks.
