+++ src/utils/index.ts
@@ -82,3 +82,19 @@ export function cachedPromise<T>(fn: () => Promise<T>): () => Promise<T> {
 
 export const readFileAsync = fsP.readFile;
 export const writeFileAsync = fsP.writeFile;
+
+export function errorMsgTag(template: TemplateStringsArray, ...substitutions: unknown[]): string {
+    return template
+        .map((str, index) =>
+            index === 0
+                ? str
+                : (
+                    inspect(substitutions[index - 1], {
+                        depth: 0,
+                        breakLength: Infinity,
+                        maxArrayLength: 5,
+                    })
+                ) + str
+        )
+        .join('');
+}
